AMI-1761 sidecar should ABEND if a failure to register occurs.
nb: this ABEND only happens if eureka cannot be reached for the initial
registration.  Because it is an abnormal exit, supervisor will by default
restart the process after a short period of time and it will retry.  This
should be sufficient to solve the issue raised in AMI-1761.

AMI-2067 bump version

AMI-2067 abort after eureka error, delay on startup to spoof supervisord